Adapting tests to squants

+++ src/test/groovy/edu/ie3/simona/model/participant/HpModelTest.groovy
@@ -137,7 +137,7 @@ class HpModelTest extends Specification {
     def nextInnerTemperature = hpModel.calculateNextState(hpData).innerTemperature()
 
     then:
-    Math.abs((nextInnerTemperature.toCelsiusScale() - expectedNewInnerTemperature)) < 0.0001
+    Math.abs((nextInnerTemperature.toCelsiusScale().doubleValue() - expectedNewInnerTemperature.doubleValue())) < 0.0001
 
     where:
     hpState                                                                                     || expectedNewInnerTemperature // (isRunning, tooHigh, tooLow)
